requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
clearml==1.16.5
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
einops==0.8.0
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
furl==2.1.3
gradio==5.4.0
gradio_client==1.4.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
orderedmultidict==1.0.1
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pathlib2==2.3.7.post1
pillow==11.0.0
pluggy==1.5.0
portalocker==2.10.1
psutil==5.9.8
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
ruff==0.7.1
safehttpx==0.1.1
safetensors==0.4.5
-e git+https://github.com/facebookresearch/segment-anything-2@c2ec8e14a185632b0a5d8b161928ceb50197eddc#egg=SAM_2
samv2==0.0.4
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
spaces==0.30.4
starlette==0.41.2
supervision==0.24.0
sympy==1.13.1
timm==1.0.11
tokenizers==0.20.1
tomlkit==0.12.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.6
transformers==4.46.0
triton==3.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
